sesion_08/requirements.txt
# Django
Django>=3.1.2,<3.2.0

# Django REST Framework
djangorestframework>=3.12.1,<3.13.0
Markdown>=3.3.2,<3.4.0
django-filter>=2.4.0,<2.5.0

# JWT
djangorestframework-simplejwt>=4.4.0,<4.5.0

# Testing
coverage>=5.3,<5.4
